meta: New dotfiles binary installation
 - Added softwareupdate and forced install if script is not running from path
 - Run homebrew on mac before anything else, to ensure proper git and
 gcc binaries
 - Move install and post-install to scripts folder
 - Remove vim.plug from repo. This way, our autoload from .vimrc will work as intended
 - Add Makefile with a few instructions, use by the dotfiles bin
 - Update installation script to use pip2
 - Update realpath check: ZSH has no realpth builtin, so we use
 ${PATH:A} instead
 - Update link script to not overwrite existing links
 - Exit when error on cloning dotfiles
 - Change protocol for cloning https instead of git://
 - Make sure the first run continues on the prompt instead of the piped zsh
 - Test if Terminal is killable after running macos.zsh options script
 - Remove unused flagging

+++ scripts/link-git-hooks.zsh
@@ -8,8 +8,10 @@ function link_git_hooks() {
   local git_hooks=($(find "${DOTFILES}/config/git/.git_template/hooks" -type f))
   for hook in $git_hooks; do
     local final_name=$HOME/${hook##$DOTFILES/config/git/}
-    e_info "Linking $(basename $hook)"
-    ln -sf $hook $final_name
+    if [[ ! -e $final_name ]]; then
+        e_info "Linking $(basename $hook)"
+        ln -sf $hook $final_name
+    fi
   done
 }
